+++ src/add-ons/kernel/file_systems/btrfs/Attribute.h
@@ -39,7 +39,7 @@ public:
 private:
 			status_t			_Lookup(const char* name, size_t nameLength,
 									btrfs_dir_entry** entries = NULL,
-									size_t* length = NULL);
+									uint32* length = NULL);
 			status_t			_FindEntry(btrfs_dir_entry* entries,
 									size_t length, const char* name,
 									size_t nameLength,
